requirements.txt
python-dotenv
langchain
pypdf
ipykernel
langchain-community
langchain-huggingface
langchain-groq
FAISS-cpu
streamlit
langchain-core
langchain-groq
sentence-transformers
-e . ##looks for setup.py in the current directory